refactor: restructure compositionRoot by folders
