+++ src/api/infrastructure/compositionRoot/controllers/subscriptionController.js
@@ -1,0 +1,6 @@
+const SubscriptionController = require('../../../presentation/controllers/subscription/SubscriptionController');
+const subscriptionService = require('../services/subscriptionService');
+
+const subscriptionController = new SubscriptionController(subscriptionService);
+
+module.exports = subscriptionController;
