--- src/api/infrastructure/compositionRoot/subscriptionController.js
@@ -1,6 +1,0 @@
-const SubscriptionController = require('../../presentation/controllers/subscription/SubscriptionController');
-const subscriptionService = require('./subscriptionService');
-
-const subscriptionController = new SubscriptionController(subscriptionService);
-
-module.exports = subscriptionController;
