+++ src/api/infrastructure/compositionRoot/validators/cityValidator.js
@@ -1,6 +1,6 @@
-const WeatherAPICityValidator = require('../validation/WeatherAPICityValidator');
-const OpenWeatherCityValidator = require('../validation/OpenWeatherCityValidator');
-const CityValidationChain = require('../../application/validation/CityValidationChain');
+const WeatherAPICityValidator = require('../../validation/WeatherAPICityValidator');
+const OpenWeatherCityValidator = require('../../validation/OpenWeatherCityValidator');
+const CityValidationChain = require('../../../application/validation/CityValidationChain');
 
 const createValidator = () => {
   const weatherAPICityValidator = new WeatherAPICityValidator();
