+++ src/api/infrastructure/compositionRoot/validators/validateSubscriptionInput.js
@@ -1,8 +1,8 @@
 const {
   createValidateSubscriptionInput,
-} = require('../../presentation/middlewares/createValidateSubscriptionInput');
+} = require('../../../presentation/middlewares/createValidateSubscriptionInput');
 const { createValidator } = require('./cityValidator');
-const { isValidEmail } = require('../validation/validateSubscriptionFields');
+const { isValidEmail } = require('../../validation/validateSubscriptionFields');
 
 const validateSubscriptionInput = createValidateSubscriptionInput(
   createValidator(),
