+++ src/api/infrastructure/compositionRoot/validators/validateWeatherInput.js
@@ -1,6 +1,6 @@
 const {
   createValidateWeatherInput,
-} = require('../../presentation/middlewares/createValidateWeatherInput');
+} = require('../../../presentation/middlewares/createValidateWeatherInput');
 const { createValidator } = require('./cityValidator');
 
 const validateWeatherInput = createValidateWeatherInput(createValidator());
